refactor(amazon): de-angularize ScalingPolicyWriter
