+++ app/scripts/modules/amazon/src/serverGroup/index.ts
@@ -1,4 +1,4 @@
 export * from './configure';
 export * from './details/scalingPolicy/ScalingPolicyTypeRegistry';
-export * from './details/scalingPolicy/scalingPolicy.write.service';
+export * from './details/scalingPolicy/ScalingPolicyWriter';
 export * from './details/sections';
