+++ app/scripts/modules/titus/src/serverGroup/details/scalingPolicy/alarmBasedSummary.component.js
@@ -4,15 +4,12 @@ const angular = require('angular');
 
 import { AccountService, CONFIRMATION_MODAL_SERVICE, TaskExecutor } from '@spinnaker/core';
 
-import { SCALING_POLICY_WRITE_SERVICE } from '@spinnaker/amazon';
-
 import './scalingPolicySummary.component.less';
 
 module.exports = angular
   .module('spinnaker.titus.serverGroup.details.scalingPolicy.alarmBasedSummary.component', [
     require('./upsert/upsertScalingPolicy.controller').name,
     CONFIRMATION_MODAL_SERVICE,
-    SCALING_POLICY_WRITE_SERVICE,
   ])
   .component('titusAlarmBasedSummary', {
     bindings: {
@@ -21,7 +18,7 @@ module.exports = angular
       application: '=',
     },
     templateUrl: require('./alarmBasedSummary.component.html'),
-    controller: function($uibModal, scalingPolicyWriter, confirmationModalService) {
+    controller: function($uibModal, confirmationModalService) {
       this.$onInit = () => {
         AccountService.getAccountDetails(this.serverGroup.account).then(details => {
           // alarmServerGroup is used to trick the chart rendering into using AWS metrics
