+++ app/scripts/modules/titus/src/serverGroup/details/scalingPolicy/targetTracking/upsertTargetTracking.controller.ts
@@ -42,7 +42,6 @@ export class UpsertTargetTrackingController implements IComponentController {
 
   constructor(
     private $uibModalInstance: IModalServiceInstance,
-    private scalingPolicyWriter: ScalingPolicyWriter,
     public policy: ITitusTargetTrackingPolicy,
     public serverGroup: ITitusServerGroup,
     public alarmServerGroup: IAlarmRenderingServerGroup,
@@ -74,7 +73,7 @@ export class UpsertTargetTrackingController implements IComponentController {
       application: this.application,
       title: `${action} scaling policy for ${this.serverGroup.name}`,
       modalInstance: this.$uibModalInstance,
-      submitMethod: () => this.scalingPolicyWriter.upsertScalingPolicy(this.application, command),
+      submitMethod: () => ScalingPolicyWriter.upsertScalingPolicy(this.application, command),
     });
 
     this.taskMonitor.submit();
